Fully rename kiosks to devices
