+++ apps/client/src/components/devices/delete-dialog.tsx
