+++ apps/client/src/components/devices/update-dialog.tsx
